requirements.txt
mkdocs==1.5.3
mkdocs-material==9.4.6
mkdocs-mermaid2-plugin==1.1.1
mkdocs-git-revision-date-plugin==0.3.2
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-macros-plugin==1.0.4
mkdocs-minify-plugin==0.7.1
simple-salesforce==1.12.5
jinja2==3.1.2
python-dotenv==1.0.0
